requirements.txt
Django==3.2.6
djangorestframework==3.14.0
django-oauth-toolkit==2.3.0
django-cors-headers==4.0.0
psycopg2==2.9.7
dj-database-url==0.5.0
gunicorn==21.2.0
nltk==3.5
numpy==1.21.5
matplotlib==3.8.1
scikit-learn==1.3.2
